requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pytz==2024.1
flask-mail==0.10.0
python-dotenv==1.0.0

